requirements.txt
comtypes
pillow